refactor: CollectionListRouteType 위치 변경
