+++ app/src/main/java/com/flint/presentation/collectiondetail/navigation/CollectionDetailNavigation.kt
@@ -6,7 +6,7 @@ import androidx.navigation.NavGraphBuilder
 import androidx.navigation.NavOptions
 import androidx.navigation.compose.composable
 import com.flint.core.navigation.Route
-import com.flint.domain.type.CollectionListRouteType
+import com.flint.core.navigation.model.CollectionListRouteType
 import com.flint.presentation.collectiondetail.CollectionDetailRoute
 
 fun NavController.navigateToCollectionDetail(
