+++ app/src/main/java/com/flint/presentation/collectionlist/navigation/CollectionListNavigation.kt
@@ -6,7 +6,7 @@ import androidx.navigation.NavGraphBuilder
 import androidx.navigation.NavOptions
 import androidx.navigation.compose.composable
 import com.flint.core.navigation.Route
-import com.flint.domain.type.CollectionListRouteType
+import com.flint.core.navigation.model.CollectionListRouteType
 import com.flint.presentation.collectionlist.CollectionListRoute
 
 fun NavController.navigateToCollectionList(
