+++ app/src/main/java/com/flint/presentation/main/MainNavigator.kt
@@ -11,7 +11,7 @@ import androidx.navigation.NavOptions
 import androidx.navigation.compose.rememberNavController
 import androidx.navigation.navOptions
 import com.flint.core.navigation.MainTabRoute
-import com.flint.domain.type.CollectionListRouteType
+import com.flint.core.navigation.model.CollectionListRouteType
 import com.flint.presentation.collectioncreate.navigation.navigateToCollectionCreate
 import com.flint.presentation.collectiondetail.navigation.navigateToCollectionDetail
 import com.flint.presentation.collectionlist.navigation.navigateToCollectionList
